+++ netpyne/simFuncs.py
@@ -1088,13 +1088,13 @@ def gatherData ():
                         elif key not in singleNodeVecs:
                             sim.allSimData[key].update(val)           # update simData dicts which are not Vectors
 
-            if len(sim.allSimData['spkt']) > 0:
-                sim.allSimData['spkt'], sim.allSimData['spkid'] = zip(*sorted(zip(sim.allSimData['spkt'], sim.allSimData['spkid']))) # sort spks
+                if len(sim.allSimData['spkt']) > 0:
+                    sim.allSimData['spkt'], sim.allSimData['spkid'] = zip(*sorted(zip(sim.allSimData['spkt'], sim.allSimData['spkid']))) # sort spks
 
-            sim.net.allPops = ODict() # pops
-            for popLabel,pop in sim.net.pops.iteritems(): sim.net.allPops[popLabel] = pop.__getstate__() # can't use dict comprehension for OrderedDict
+                sim.net.allPops = ODict() # pops
+                for popLabel,pop in sim.net.pops.iteritems(): sim.net.allPops[popLabel] = pop.__getstate__() # can't use dict comprehension for OrderedDict
 
-            sim.net.allCells = [c.__dict__ for c in sim.net.cells]
+                sim.net.allCells = [c.__dict__ for c in sim.net.cells]
 
         # gather cells, pops and sim data
         else: