+++ CHANGES.md
@@ -1,3 +1,11 @@
+# Version 0.7.5
+
+- Added methods to rename netParams dict keys in a nested format (needed for GUI)
+
+- Fix issue to make convergence+divergence connections randomization more robust and efficient (issue #254)
+
+- Fix bug when using sim.gatherOnlySimData
+
 # Version 0.7.4
 
 - Added polarity param to analysis.plotEPSPAmp() 
@@ -13,6 +21,7 @@
 - Fixed bug use int as argument to np.full in fixedInterval variable
 
 - Fixed bug that removed previously existing element from path during importCellParams()
+
 
 # Version 0.7.3